Instead of using a fixed audio buffer, call a callback for all audio rendered, and another when rendering is complete.
Use this new system to output audio using either miniaudio or by writing to a wav file based on which callback is set on the toSpeak context.
This seems to change audio output a bit, stretching some phonemes too long but generally works well.
The phoneme stretching should be researched